gcp_hosting/app_engine/requirements.txt
flask>=1.1.0
google-cloud-storage